+++ Sources/Main/Core/Geo/GeoManager.swift
@@ -1,6 +1,5 @@
 import Foundation
 import CoreLocation
-import UIKit.UIDevice
 
 /// Completion block for request location.
 ///
@@ -38,57 +37,19 @@ protocol GeoTrackable {
 // MARK: - GeoManager
 /// The object that you use to start, stop and request the delivery of location-related events to your app.
 public final class GeoManager {
-    /// The Geo Tracker.
-    ///
-    /// - Note: The Geo Tracker instantiation returns nil when the endpoint URL is not configured (`RATEndpoint`).
-    private let geoTracker: Tracker?
-
-    /// The device identifier handler.
-    private let deviceIdentifierHandler: DeviceIdentifierHandler
-    
-    private let geoSharedPreferenceHelper: GeoConfigurationHelper
-
-    /// The current location of the user.
-    private var location: CLLocation?
-
     /// The location collection configuration.
     private var configuration: Configuration?
-
+    private let geoSharedPreferenceHelper: GeoConfigurationHelper
     /// - Returns: The shared instance of `GeoManager` object.
     public static let shared: GeoManager = {
-        let device = UIDevice.current
         let dependenciesContainer = SimpleDependenciesContainer()
-        
-        guard let databaseConfiguration = DatabaseConfigurationHandler.create(databaseName: GeoTrackerConstants.databaseName,
-                                                                              tableName: GeoTrackerConstants.tableName,
-                                                                              databaseParentDirectory: Bundle.main.databaseParentDirectory) else {
-            RLogger.error(message: "The GeoTracker could not be created because the SQLite connection failed.")
-            return GeoManager(geoTracker: nil,
-                              device: device,
-                              userStorageHandler: dependenciesContainer.userStorageHandler)
-        }
-        let geoTracker = GeoTracker(dependenciesContainer: dependenciesContainer,
-                                    databaseConfiguration: databaseConfiguration)
-        return GeoManager(geoTracker: geoTracker,
-                          device: device,
-                          userStorageHandler: dependenciesContainer.userStorageHandler)
+        return GeoManager(userStorageHandler: dependenciesContainer.userStorageHandler)
     }()
 
     /// Creates a new instance of GeoManager.
     ///
-    /// - Parameter geoTracker: The GeoTracker instance or nil when the SQLite connection fails to create the GeoTracker Database.
-    /// If the GeoTracker could not be created, the other features of GeoManager are still running.
-    ///
-    /// - Parameter device: The device capability used to calculate the `ckp`.
-    ///
-    init(geoTracker: Tracker?,
-         device: DeviceCapability,
-         userStorageHandler: UserStorageHandleable) {
-        
-        self.geoTracker = geoTracker
-
-        self.deviceIdentifierHandler = DeviceIdentifierHandler(device: device,
-                                                               hasher: SecureHasher())
+    /// - Parameter userStorageHandler: Parameter of type `UserStorageHandleable` provides an interface to the user’s defaults database, where you store key-value pairs persistently across launches of your app.
+    init(userStorageHandler: UserStorageHandleable) {
         self.geoSharedPreferenceHelper = GeoConfigurationHelper(userStorageHandler: userStorageHandler)
     }
 }
@@ -103,24 +64,6 @@ extension GeoManager: GeoTrackable {
         }
         
         self.configuration = configuration
-
-        // Note: GeoManager has to calculate the location.
-        guard let location = location else {
-            return
-        }
-
-        // Note: GeoManager has to fill these values:
-        // - isAction
-        // - actionParameters
-        let state = RAnalyticsState(sessionIdentifier: Session.cks(),
-                                    deviceIdentifier: deviceIdentifierHandler.ckp())
-        state.lastKnownLocation = LocationModel(location: location,
-                                                isAction: false,
-                                                actionParameters: nil)
-
-        _ = geoTracker?.process(event: RAnalyticsEvent(name: RAnalyticsEvent.Name.geoLocation,
-                                                       parameters: nil),
-                                state: state)
     }
 
     public func stopLocationCollection() {